Added better bandwidth calculation to global calibration.
Previously used Silverman factor for all, but with extremely high statistics
this was leading to jagged edges in the pdf.

The bandwidth is now limited to a minimum of 0.1. This means that the majority
of the global calibrations use 0.1, whereas the pixel calibrations will
continue to use the calculated Silverman factor.