Altered the global KDE method to fit sources not peaks.
Splitting the TOT spectra into separate peaks can cause problems in
the pixel by pixel calibration (because the response vaies). It's
better not to split the spectra, but instead to find all peaks
together. That's what this new method does for the global calibration.

In calibrateKDEGlobal.py:
- peak option changed to source
- find the peaks by considering the derivative. If more peaks are found
than expected, the highest peaks are chosen.
- the uncertainty calculation is as before, but now happens per peak
- the details of each peak is shown on the plot
- the text file saves all peaks

In fitGlobalSurrogate.py:
- rejig the loading of numbers saved in the files